apps/data-backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==23.0.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.10.5
langchain-openai==0.3.1
langfuse==2.57.12
python-dotenv==1.0.1
python-json-logger==3.2.1
structlog==25.1.0
umap-learn==0.5.7
pandas==2.2.3
scikit-learn==1.6.1
coralogix-logger==2.0.6